aelfgpt/pyproject.toml
[tool.poetry]
name = "aelf-gpt"
version = "0.1.0"
description = ""
authors = ["Ademola Balogun <ademolabalogun91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
chainlit = "^1.1.306"
llama-index = "^0.10.55"
llama-index-llms-ollama = "^0.1.6"
chromadb = "^0.5.4"
llama-index-embeddings-huggingface = "^0.2.2"
llama-index-readers-json = "^0.1.5"
llama-index-vector-stores-chroma = "^0.1.10"
llama-index-vector-stores-mongodb = "^0.1.7"
pysqlite3-binary = "^0.5.3"
pymongo = {extras = ["srv"], version = "^4.7.2"}
aelf-sdk = "^1.2.3"
alembic = "^1.13.2"
asyncpg = {version = "^0.29.0", extras = ["sa"]}
sqlalchemy = {version = "2.0.18", extras = ["asyncio"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
